feat: Optimización y separación del detalle-alumno respecto a modal
