+++ src/app/components/detalle-alumno-dialog/detalle-alumno-dialog.component.css
